+++ tests/test_patches.py
@@ -207,6 +207,12 @@ def test_protocol_parsers():
             },
         ),
         (
+            ClientCreator._load_service_endpoints_ruleset,
+            {
+                'a60dc1f2df973c746f8787e127f3bb9308b456a9',
+            },
+        ),
+        (
             ClientCreator._register_endpoint_discovery,
             {
                 '483c6c8e035810d1b76110fc1956de76943c2f18',